+++ R/descending.R
@@ -6,7 +6,7 @@
 #' @param f a function to modify.
 #'
 #' @return A function returning a numeric vector which, if passed to
-#' \code{\link{order}}, will be used to sort the original data.
+#' \code{\link{order}}, will be used to sort some data.
 #'
 #' @examples
 #' desc_abs <- descending(abs)