+++ NEWS.md
@@ -1,14 +1,23 @@
-# Package 'funprog'
+# funprog 0.2.0
+
+## New functions
+
+* `unique_by` : remove duplicate elements, given a transformation
+* `partition` : split a vector or a list in 2 groups, given a predicate function
+
+## Improvements
+
+* `group_if` : redesigned to be considerably faster
 
 
-## funprog 0.1.0 (2019-10)
+# funprog 0.1.0
 
-### Main functions
+## Main functions
 
 * `group_if`
 * `sort_by`
 
-### Helper functions
+## Helper functions
 
 * `%on%`
 * `descending`